Seperated database handling in matching engine to occur when the Order Executor confirms it's database handling is complete. Now expected order execution to complete minimal operations on the pair of orders before returning to it, to prevent matching engine delay.
Updated matching engine to use bulkification of updates.

Updated stock order database to now delete stock orders automatically if their quanities are updated to 0.

Updated update methods in database access to now clear the update list on a stock order. Need to check for edge cases on this.

Changed any variable named errors to errorList, to prevent conflict with the errors package.